+++ tests/test_commands.py
@@ -37,7 +37,8 @@ class CommandTest(unittest.TestCase):
             'SimplePayment-prepaid',
             'SimplePayment-prepaid-direct',
             'SimplePayment-postpaid',
-            'SimplePayment-postpaid-direct'
+            'SimplePayment-postpaid-direct',
+            'SmartJudge-FairSwap'
         ]))
 
     def test_list_data_providers(self):